Made Filter Activity
Moved distance and search filter settings into a new activity
