+++ .idea/misc.xml
@@ -6,6 +6,7 @@
         <entry key="..\:/Users/alan_/Documents/IAT359/Stumble/app/src/main/res/layout/activity_desserts_and_drinks.xml" value="0.27447916666666666" />
         <entry key="..\:/Users/alan_/Documents/IAT359/Stumble/app/src/main/res/layout/activity_dining.xml" value="0.27447916666666666" />
         <entry key="..\:/Users/alan_/Documents/IAT359/Stumble/app/src/main/res/layout/activity_events.xml" value="0.27447916666666666" />
+        <entry key="..\:/Users/alan_/Documents/IAT359/Stumble/app/src/main/res/layout/activity_filter.xml" value="0.264" />
         <entry key="..\:/Users/alan_/Documents/IAT359/Stumble/app/src/main/res/layout/activity_main.xml" value="0.2" />
         <entry key="..\:/Users/alan_/Documents/IAT359/Stumble/app/src/main/res/layout/activity_saved.xml" value="0.264" />
         <entry key="..\:/Users/alan_/Documents/IAT359/Stumble/app/src/main/res/layout/activity_selected.xml" value="0.20846905537459284" />
