+++ app/src/main/AndroidManifest.xml
@@ -20,6 +20,9 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
+            android:name=".FilterActivity"
+            android:exported="false" />
+        <activity
             android:name=".SelectedActivity"
             android:exported="false" />
         <activity
